refactor(inference): lift the oriented sRGB decode into img_utils
`decode_oriented_srgb` applies EXIF orientation, flattens alpha, and converts
an embedded ICC profile to sRGB. It has lived in `ltx2/preprocess.rs` since
that module needed upstream-parity still conditioning, and its doc even says
the module is LTX-2 specific on purpose.

#1222 needs the same upright decode for identity photographs, and the wrong
way to get it is a second orientation table. Which way up a photograph goes is
not a per-family question, so the function moves to `img_utils` — the crate's
existing shared image module — and `ltx2::preprocess` re-exports it under its
original name so every LTX-2 call site, test, and upstream citation reads
exactly as before.

`convert_icc_to_srgb` travels with it, since it is the second half of the same
operation. Its coverage stays in the LTX-2 tests, which is where the
upstream-parity requirement for the assume-sRGB fallback actually lives.

Pure move: no behaviour change, and the LTX-2 preprocess tests pass unchanged.

Refs #1222
