+++ crates/mold-inference/src/img_utils.rs
@@ -1,8 +1,9 @@
 //! Image decoding and preprocessing utilities for img2img.
 
-use anyhow::{bail, Result};
+use anyhow::{bail, Context, Result};
 use candle_core::{DType, Device, Tensor};
-use image::{DynamicImage, RgbImage};
+use image::{DynamicImage, ImageDecoder, RgbImage};
+use std::io::Cursor;
 
 /// Normalization range for source images before VAE encoding.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
@@ -151,6 +152,112 @@ pub fn decode_mask_image(
     let tensor = tensor.to_dtype(dtype)?.to_device(device)?;
 
     Ok(tensor)
+}
+
+// ---------------------------------------------------------------------------
+// Oriented sRGB decode
+//
+// Shared by LTX-2 still conditioning (`ltx2::preprocess`, which re-exports it)
+// and PuLID identity extraction (`identity`). One orientation path in the
+// crate on purpose: a second table is a second chance to disagree about which
+// way up a photograph is.
+// ---------------------------------------------------------------------------
+
+/// Decode image bytes into an upright sRGB image: EXIF orientation is
+/// applied, alpha is flattened, and an embedded ICC profile is converted
+/// to sRGB (malformed profiles warn and assume sRGB, mirroring upstream
+/// decode.py:164-166).
+pub fn decode_oriented_srgb(bytes: &[u8]) -> Result<RgbImage> {
+    let reader = image::ImageReader::new(Cursor::new(bytes))
+        .with_guessed_format()
+        .context("failed to sniff source image format")?;
+    let mut decoder = reader
+        .into_decoder()
+        .context("failed to decode source image")?;
+    let orientation = decoder
+        .orientation()
+        .unwrap_or(image::metadata::Orientation::NoTransforms);
+    let icc = decoder.icc_profile().unwrap_or_default();
+    let mut decoded =
+        DynamicImage::from_decoder(decoder).context("failed to decode source image")?;
+    // Upstream order (decode.py:143-163): orient, flatten to RGB, then
+    // convert color to sRGB.
+    decoded.apply_orientation(orientation);
+    // A grayscale source with a Gray ICC profile must be transformed in
+    // its own layout — moxcms (correctly) refuses an Rgb-layout transform
+    // from a Gray profile, and pre-flattening to RGB would turn the
+    // conversion into the assume-sRGB fallback (codex review, PR #1072).
+    let is_grayscale = matches!(
+        decoded.color(),
+        image::ColorType::L8
+            | image::ColorType::La8
+            | image::ColorType::L16
+            | image::ColorType::La16
+    );
+    let rgb = decoded.to_rgb8();
+    Ok(match icc {
+        Some(profile) if !profile.is_empty() => {
+            let layout = if is_grayscale {
+                moxcms::Layout::Gray
+            } else {
+                moxcms::Layout::Rgb
+            };
+            convert_icc_to_srgb(rgb, &profile, layout)
+        }
+        _ => rgb,
+    })
+}
+
+/// Convert `rgb` from the embedded `profile` to sRGB, reading the source
+/// pixels in `source_layout` (`Gray` collapses the flattened RGB back to
+/// one channel — the three are identical for a decoded grayscale image).
+/// Any failure — malformed profile, unsupported connection — logs and
+/// returns the pixels unchanged (assume-sRGB), exactly upstream's
+/// fallback.
+pub(crate) fn convert_icc_to_srgb(
+    rgb: RgbImage,
+    profile: &[u8],
+    source_layout: moxcms::Layout,
+) -> RgbImage {
+    let source = match moxcms::ColorProfile::new_from_slice(profile) {
+        Ok(profile) => profile,
+        Err(error) => {
+            tracing::warn!("ignoring malformed embedded ICC profile: {error}");
+            return rgb;
+        }
+    };
+    let srgb = moxcms::ColorProfile::new_srgb();
+    let options = moxcms::TransformOptions {
+        rendering_intent: moxcms::RenderingIntent::Perceptual,
+        ..Default::default()
+    };
+    let transform =
+        match source.create_transform_8bit(source_layout, &srgb, moxcms::Layout::Rgb, options) {
+            Ok(transform) => transform,
+            Err(error) => {
+                tracing::warn!("cannot convert embedded ICC profile to sRGB: {error}");
+                return rgb;
+            }
+        };
+    let (width, height) = rgb.dimensions();
+    let src_rgb = rgb.into_raw();
+    let src: std::borrow::Cow<'_, [u8]> = match source_layout {
+        moxcms::Layout::Gray => src_rgb
+            .as_chunks::<3>()
+            .0
+            .iter()
+            .map(|px| px[0])
+            .collect::<Vec<_>>()
+            .into(),
+        _ => (&src_rgb).into(),
+    };
+    let mut dst = vec![0u8; width as usize * height as usize * 3];
+    if let Err(error) = transform.transform(&src, &mut dst) {
+        tracing::warn!("ICC-to-sRGB conversion failed: {error}");
+        return RgbImage::from_raw(width, height, src_rgb)
+            .expect("source buffer matches dimensions");
+    }
+    RgbImage::from_raw(width, height, dst).expect("destination buffer matches dimensions")
 }
 
 #[cfg(test)]
